Implemented Optional Drag Drop Fix
Previously, if you were on Windows and dragged a .pac onto the executable from another folder, the program would fail to run. This was because dragging a file from another folder makes that other folder the program's "working directory", which means any relative paths are treated as relative to that folder rather than the folder the executable is in. As a result, when the program would go to look for the "patches.txt" file or process relative patch paths specified within "patches.txt", it would fail to find anything unless you'd copied your "patches.txt" and any relative patches to that directory.

This struck me as silly, so I've implemented an admittedly sorta hacky fix. Now, if you #define WIN_DRAG_DROP_FIX, the program will force "patches.txt" and any relative paths it specifies to be treated as relative to the executable folder, enabling the ability to drag .pacs from other folders and process them.

Note, though, if you're working in Visual Studio with this enabled, you'll need to move your "patches.txt" into the "Debug" folder, since that's where the executable will be. Same for any relative paths defined in it. Note that you can simply leave WIN_DRAG_DROP_FIX undefined while debugging and editing the program, then define it when compiling a version you intend to actively use or release.