Edited to Ensure Applying This Doesn't Change Any Values
The test patches are meant purely to demonstrate that the relevant actions are functioning as intended, they are not meant to change anything about files they are applied to. This one was causing a change (doubled jumpsquat and jump height), it has been edited so that it no longer does.